+++ iBMSC/Utilities.vb
@@ -24,7 +24,7 @@ Namespace Editor
             End If
             Return 0
         End Function
-        Public Function C10to36(ByVal xStart As Integer) As String
+        Public Function C10to36(ByVal xStart As Long) As String
             If xStart < 1 Then xStart = 1
             If xStart > 1295 Then xStart = 1295
             Return C10to36S(xStart \ 36) & C10to36S(xStart Mod 36)
